refactor: group admin commands under `bench admin` namespace
Move the control-plane commands and upgrade into pilot/commands/admin/ under a
new `admin` group, dropping the now-redundant `-admin`/`admin` suffixes:

  build-admin            -> admin build
  enroll                 -> admin enroll
  generate-admin-session -> admin generate-session
  issue-site-token       -> admin issue-site-token
  set-central-config     -> admin set-central-config
  upgrade                -> admin upgrade

Clean break, no aliases (no external callers; pre-alpha). Update the discovery
baseline and docs.
